Implementing time-dependency of physical params
While doing this, the isssue of duplicate reactions (with different time ranges) have to be dealt with.

The applicable temperature range of each reaction should also be properly handled, to avoid jumps in the reaction rate as the temperature is changed.

The time-dependency of physical parameters are implemented with user-input data tables, hence interpolation is needed (currently linear interpolation is used).